Use GetCurrentUser to set runAsSUer for ds check
The current runAsUser := 1000 creates issues when running in OCP.
If we assume that the user is set to none root in our checks
we can just grab that user ID and use that in the DS that is created.

This will also work with runAsUser set in a standard k8s env or just
running it as a container since we have set 999 as the default user in
the containers.