fix(http-check): remove password logging with basic auth
It is possible to configure basic auth credentials by passing them directly
in the URL with this format: `http://user:password@doamin.tld/path`

With such a configuration, `http-check` logs the password in many places.

This patch uses the `url.Redacted()` method to suppress the password (but not
the username) from the logs and the generated metrics.

Signed-off-by: Aloÿs Augustin <aloys.augustin@aqemia.com>